+++ src/types/nftscan-type.ts
@@ -12,6 +12,7 @@ export enum EvmChain {
   PLATON = 'platon',
   AVALANCHE = 'avalanche',
   CRONOS = 'cronos',
+  FANTOM = 'fantom',
 }
 
 /**
@@ -140,4 +141,19 @@ export interface BaseNsPaginationResData {
    * The total of the results matching the query
    */
   total: number;
+}
+
+/**
+ * The response data of refresh metadata api
+ */
+export interface RefreshMetadataResponse {
+  /**
+   * The reason for the FAIL status
+   */
+  reason: string;
+
+  /**
+   * Task status. SUCCESS means the task was successfully submitted, FAIL means the task was not submitted due to some reason
+   */
+  status: string;
 }